requirements.txt
beautifulsoup4
lxml
requests
pyTelegramBotAPI
python-decouple